refactor[#345] : 위경도로 현재 주소 가져오는 api 요청 함수에 apiClient 적용
