--- src/features/map/api/index.ts
@@ -1,1 +1,0 @@
-export * from "./location";
